requirements.txt
opencv-python-headless==4.12.0.88
face-recognition==1.3.0
deepface==0.0.95
pandas==2.3.3
numpy==2.2.6
PyQt5==5.15.11
matplotlib==3.9.4
seaborn==0.13.2
openpyxl==3.1.5
reportlab==4.2.5
Pillow==12.0.0
dlib==20.0.0
tensorflow==2.20.0
